Add prompt-based searching as commands/shortcuts
To make it easier to search for symbols that are not immediately present
in the source code without changing the enabling prompt-before-
searching, add additional commands and shortcuts to perform searches
with a prompt for typing in the symbol name to search for.

Delete the prompt_before_searching option as it is obsolete with this
change.

Fixes #8.